Update model hidden layers

+++ data.py
@@ -4,7 +4,7 @@ from torch.utils.data import Dataset
 class CustomDataset(Dataset):
     def __init__(self, data, n_atomtype=10):
         self.X = torch.tensor(data['R'], dtype=torch.float)
-        self.Z = torch.nn.functional.one_hot(torch.tensor(data['Z'].squeeze(), dtype=torch.long), num_classes=n_atomtype)
+        self.Z = torch.nn.functional.one_hot(torch.tensor(data['Z'].squeeze(), dtype=torch.long), num_classes=n_atomtype).float()
         
         n_sample = self.X.shape[0]
         n_atom = self.X.shape[1]
